tam on

+++ db/migrate/20180415064404_create_orders.rb
@@ -3,7 +3,7 @@ class CreateOrders < ActiveRecord::Migration[5.1]
     create_table :orders do |t|
       t.integer :user_id
       t.integer :guess_id
-      t.float :total_price
+      t.decimal :total_price
       t.boolean :status
       t.date :date_order
       t.date :date_receipt
